Refactored to use zf-console
Rewrote the library to use zf-console. This removes the requirement for the ZF2
MVC, and replaces it with `zfcampus/zf-console`, plus a short list of other
requirements:

- `zendframework/zend-dom` (this was techniclly required previously)
- `zendframework/zend-filter`
- `zendframework/zend-validator`

The console command functionality was moved to `bin/phly-comic.php`, which now
supports the following commands:

- `list`
- `fetch <comic> [--output=]`
- `fetch-all [--output=]`

To enable this, a `config/routes.php` was added which defines all routes, along
with any required constraints, filters, and/or validators. `FetchController` was
removed, and replaced with `Console\ListComics`, `Console\FetchComic`, and
`Console\FetchAllComics`; the latter two extend
`Console\AbstractConsoleHandler`, which provides some base functionality around
fetching individual comics, providing templates, and helper functions for
reporting success and error messages. Overall, a large number amount of code was
removed.

A `composer.json` was added, which specifies the requirements, adds autoloading,
and marks `bin/phly-comic.php` as a binary for installation.