Fixes comic discovery in Dave Kellet comics
The first item in the feed is often a promotional item, meaning the comic gets
lost. This change adds some functionality to the AbstractRssSource to allow
determining if the feed item is of interest; if not, it moves to the next.

For these comics, it checks to see if the guid includes the path segment
`/archive/`; if it doesn't, it's a news item.