Fixes fetching of Ctrl-Alt-Del
Change in CMS means change in URIs. The feed cannot be used, as it only
links to a page with a comic, but does not contain a link to the comic
image itself. The DOM-based approach works, but requires a breaking
change to the `getDailyUrl()` method to allow passing in the
`Zend\Dom\Query` instance so we can execute another query to extract the
link to the page.