fix: fixes FoxTrot source
RSS feed changed to only provide links to the comic pages, instead of
also including the image URL directly. This patch updates the FoxTrot
comic source to use an approach similar to that used by PennyArcade, and
follow the link in the feed, and then parse that page for the comic.