refactor: extract common logic of foxtrot, pennyarcade, into new base class AbstractRssAndDomSource
This patch creates `AbstractRssAndDomSource`, which extends
`AbstractRssSource` and composes `XPathTrait` in order to extract an
image from a page link provided by a feed.

It defines one additional abstract function,
`validateFeedItem(SimpleXMLElement $item) : bool`, which is used by
`AbstractRssAndDomSource` to determine if it should continue with or
ignore a given item when trying to find an image source.