feat: adds The Oatmeal as a source
To accomplish this, I needed to add an `AbstractDateOrderedRssSource`, which sorts the items by date (descending) before choosing the item to present; this is due to the fact that an item gets "pinned" to the top of this particular feed.

In doing so, I discovered that the mechanism I used to import a feed was flawed, as the process of importing via DOM and then translating to SimpleXML was stripping namespaces in many instances.
I've now switched to just `simplexml_load_string()` and checking for a non-false result.
This does not appear to have negative impact.

Finally, I added The Oatmeal as an implementation of the new comic source, without overriding any methods; it works as expected.