fix: fetching of FoxTrot
The RSS feed has links which go to a page that does UserAgent sniffing.
As such, the returned page was a 403.

Adding a Chrome user agent then led to a 301 redirect.

To make it all work, I switched to fetching the page via curl, setting the option to follow redirects, and adding a user agent.