refactor: refactor ComicFactory to be PSR-11 container
Refactors ComicFactory to be a PSR-11 container, with an added method `getSupported()`.
Refactors all console classes to consume an _instance_ instead of using it statically, and to also use Psr18Discovery to create and return the client to use with a comic source.
Refactors console classes to consume comic instances instead of aliases.